# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: Todo API with express
servers:
  - url: http://localhost:3001
paths:
  /auth/register:
    post:
      summary: 회원가입
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: test1
                email:
                  type: string
                  example: test1@test.test
                password:
                  type: string
                  example: test12008!
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: Boolean
                    example: true
                  message:
                    type: string
                    example: 회원가입 성공
                  user:
                    type: object
                    properties:
                      user_id:
                        type: int
                        example: 2
                      username:
                        type: string
                        example: test1
                      password:
                        type: string
                        example: >-
                          $10$vUB1q.9UghVtKXJVEhvXB.CQrwJNklnV92x5uQ0ddHRfZIWPus7Li
                      email:
                        type: string
                        example: test1@test.test
                      created_at:
                        type: string
                        example: '2025-12-09 10:18:00.300'
  /auth/emailCheck.js:
    get:
      summary: 이메일 중복 검사
      parameters:
        - in: query
          name: email
          description: 이메일 검사
          schema:
            type: string
            example: test@test.test
      responses:
        '200':
          description: 사용 가능한 이메일
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 사용 가능한 이메일입니다.
        '400':
          description: 이메일을 입력하지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 이메일을 입력해주세요.
        '409':
          description: 이메일이 이미 사용 중
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 이미 사용 중인 이메일입니다.
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 서버에 오류가 발생했습니다.
  /auth/login:
    post:
      summary: 로그인
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test@test.test
                password:
                  type: string
                  example: test2008!
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 로그인 성공
                  token:
                    type: string
                    example: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJ1c2VybmFtZSI6ImV4YW1wbGV1c2VyIiwiZW1haWwiOiJleGFtcGxlQGV4YW1wbGUuY29tIiwiY3JlYXRlZF9hdCI6IjIwMjEtMDEtMDEiLCJpYXQiOjE2ODkzNDU2MDAsImV4cCI6MTY4OTM0OTIwMH0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                  user:
                    type: object
                    properties:
                      user_id:
                        type: int
                        example: 1
                      username:
                        type: string
                        example: test
                      email:
                        type: string
                        example: test@test.test
                      created_at:
                        type: string
                        example: '2025-12-10'
        '500':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: Boolean
                    example: false
                  message:
                    type: string
                    example: 서버 오류
  /users/me:
    get:
      summary: 사용자의 프로필을 조회
      description: JWT 토큰을 통해 로그인한 사용자의 프로필 정보를 반환합니다.
      responses:
        '200':
          description: 성공적으로 프로필 정보를 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      user_id:
                        type: int
                        example: 1
                      username:
                        type: string
                        example: test1
                      email:
                        type: string
                        example: test1@test.test
                      created_at:
                        type: string
                        example: '2025-12-09 10:18:00.300'
        '401':
          description: 토큰이 없거나 유효하지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 토큰 없음
  /todos:
    post:
      summary: TO DO 생성
      description: 토큰과 user의 고유 id가 필요합니다.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 회의
                description:
                  type: string
                  example: 일일 보고에 관하여
                deadline:
                  type: string
                  example: '2025-11-25'
      responses:
        '200':
          description: TO DO 추가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  datas:
                    type: object
                    properties:
                      todo_id:
                        type: int
                        example: 1
                      title:
                        type: string
                        example: 회의
                      description:
                        type: string
                        example: 일일 보고에 관하여
                      date:
                        type: string
                        example: '2025-11-25'
                      created_at:
                        type: string
                        example: '2025-12-10T12:05:00'
                      user_id:
                        type: int
                        example: 1
        '401':
          descript: 토큰이 없을떄.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 토큰없음
    get:
      summary: TO DO List 조회
      description: 토큰의 user id를 받아 TO DO 스키마에 저장된 user id에 관한 모든 데이터를 불러옴
      responses:
        '200':
          description: 성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  datas:
                    type: object
                    properties:
                      todo_id:
                        type: int
                        example: '1'
                      title:
                        type: string
                        example: 회의 안건
                      description:
                        type: string
                        example: 회의 안건을 만들어 봅시다.
                      date:
                        type: string
                        example: '2025-12-10'
                      completed:
                        type: boolean
                        example: false
        '401':
          description: null
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    message: 토큰 없음
  /todos/{id}:
    delete:
      summary: TO DO 삭제
      description: 토큰의 유저 고유 id와 To Do 에 저장된 유저 고유 ID가 같으면 삭제
      parameters:
        - in: path
          name: id
          required: true
          description: 삭제할 TO DO의 ID
          schema:
            type: int
            example: 1
      responses:
        '200':
          description: 삭제성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  successs:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 삭제성공!
        '401':
          desription: 삭제실패 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 삭제 실패!
    put:
      summary: To Do 수정
      description: 토큰의 userid와 todo_id가 맞는지 확인 후 정보 수정
      parameters:
        - in: path
          name: id
          required: true
          description: 수정할 To DO 의 ID
          schema:
            type: int
            example: 1
      requestBody:
        description: 수정할 데이터
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 회의 안건
                description:
                  type: string
                  example: 회의 안건을 만들어 봅시다.
                date:
                  type: string
                  example: '2025-12-10'
      responses:
        '200':
          description: 수정 성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 수정 성공
        '401':
          description: 수정 실패 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 수정 실패
  /todos/checkOk:
    get:
      summary: TO DO List completed 조회
      description: >-
        토큰의 user id를 받아 TO DO 스키마에 저장된 user id에 관한 모든 데이터를 불러옴 단. completed 항목이
        true일 겅우만
      responses:
        '200':
          description: 성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  datas:
                    type: object
                    properties:
                      todo_id:
                        type: int
                        example: '1'
                      title:
                        type: string
                        example: 회의 안건
                      description:
                        type: string
                        example: 회의 안건을 만들어 봅시다.
                      date:
                        type: string
                        example: '2025-12-10'
                      completed:
                        type: boolean
                        example: true
        '401':
          description: null
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    message: 토큰 없음
  /todos/checkNo:
    get:
      summary: TO DO List Not completed 조회
      description: >-
        토큰의 user id를 받아 TO DO 스키마에 저장된 user id에 관한 모든 데이터를 불러옴, 단 completed가
        false인 것만
      responses:
        '200':
          description: 성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  datas:
                    type: object
                    properties:
                      todo_id:
                        type: int
                        example: '1'
                      title:
                        type: string
                        example: 회의 안건
                      description:
                        type: string
                        example: 회의 안건을 만들어 봅시다.
                      date:
                        type: string
                        example: '2025-12-10'
                      completed:
                        type: boolean
                        example: false
        '401':
          description: null
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    message: 토큰 없음
  /todos/{title}:
    get:
      summary: TO DO List title 조회
      description: >-
        토큰의 user id를 받아 TO DO 스키마에 저장된 user id에 관한 모든 데이터를 불러옴, 단 title과 비슷한 으름의
        데이터만
      parameters:
        - in: path
          name: title
          required: true
          description: 비슷한 이름의 TO Do 조회
          schema:
            type: int
            example: 회의
      responses:
        '200':
          description: 성공 시
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  datas:
                    type: object
                    properties:
                      todo_id:
                        type: int
                        example: '1'
                      title:
                        type: string
                        example: 회의 안건
                      description:
                        type: string
                        example: 회의 안건을 만들어 봅시다.
                      date:
                        type: string
                        example: '2025-12-10'
                      completed:
                        type: boolean
                        example: false
        '401':
          description: null
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    message: 토큰 없음
components: {}
tags: []